Fix crash of odb/test/test in a freshly cloned system.
The edb.so plugin #includes util.hh, which declares ParserMeta's
constructor. The implementation of this constructor is in util.cc,
but edb.so wasn't linked against libedbutil.so, so the reference
to the ParserMeta constructor was unresolved in the plugin.

The attempt to:

    create function ... soname 'edb.so'

complained

    Can't open shared library 'edb.so' (errno: 0 ...)

The "fix" attempted here is to link edb.so against -ledbutil.

While this seems to allow obj/test/test to work, it's
surprising that:

> mysql -u root -pletmein mysql
> select count(*) from func;

now returns 0.  Previously, there were UDFs in the mysql.func table.
So perhaps the problem isn't completely fixed by this.

---

ALSO: changed the make install rule for edb.so to change the
owner and group to mysql.  NOTE - this requires that make install
be run as root (sudo make install).